add some casts for VS compatibility

+++ trng/binomial_dist.hpp
@@ -105,7 +105,8 @@ namespace trng {
     // random numbers
     template<typename R>
     int operator()(R &r) {
-      return utility::discrete(utility::uniformoo<double>(r), P.P_.begin(), P.P_.end());
+      return static_cast<int>(
+          utility::discrete(utility::uniformoo<double>(r), P.P_.begin(), P.P_.end()));
     }
     template<typename R>
     int operator()(R &r, const param_type &P) {
